refactored the current widget
